+++ CHANGELOG.md
@@ -26,6 +26,8 @@
 * added connection argument to concerto.table.query and concerto.table.lastInsertId 
 * clicking outside of modal in panel no longer closes modal
 * increased  data table's data grid columns min width
+* added flowIndex argument to c.get and c.set
+* replaced posOffset argument with flowIndexOffset in c.get and c.set
 
 ## 5.0.10 (2020-01-13)